Move `TestBlockProvider` to the test file as it is not used anywhere else.
